chore: Clean-up before to merge the PR
