+++ crates/weaver_forge/src/extensions/mod.rs
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 //! Custom filters used by the template engine.
-pub mod acronym;
 pub mod ansi;
 pub mod case;
 pub mod code;
 pub mod otel;
+pub mod util;
